Removes almost all @objc annotations
The ones that have been left intact are necessary to spin up an OBA app from an Objective-C app delegate. I'll selectively re-annotate the codebase with @objc as needed in order to make anything work from Objective-C that needs to, but I don't think that annotating absolutely everything makes sense.