DRYs up app plist files
This commit takes us back to the state we were in a couple weeks ago where applications' Info.plists are no longer checked in, and the critical information they contain is embedded in the project.yml files instead. I think this is the last time I vacillate on this. Here's why:

1. We won't see the version thrashing on the Info.plist files that has been such a pain of late, especially on rebases.
2. This allows de-duplication of apps' Info.plist files, which is a huge win in terms of ensuring that config information is shared between all of the apps that will be built on top of this repo.
3. This should make it easier to deal with different InfoPlist.strings data in different app targets. More to come on that soon.