Adds proper documentation using Jazzy
I had removed Jazzy a while back because it needed more work in order to work right, and now I'm finally ready to commit to it.

This commit adds support for generating documentation for both OBAKit and OBAKitCore using Jazzy.

This commit also:

* Rearranges several files
* Modifies the visibility of many symbols in the two frameworks
* Adds some more doc comments

Lots more to do, but this feels like a good start!