Use a JSON body in the PaymentIntent creation API instead of a normal HTTP POST body
I was running into a bug with our code that created an HTTP POST body from a dictionary. Specifically, it was removing `+` characters from email addresses. This should get fixed, but it was easier in the moment to simply encode the request body as JSON and decode it accordingly on the server side.