Split Stripe SDK into optional product feature
Fixes https://github.com/OneBusAway/onebusaway-ios/issues/705 - Stripe SDK should not be included in white label apps that do not accept donations

Similar to the Firebase and OneSignal SDKs, Stripe is now isolated from apps that do not link in that framework. Unfortunately, I haven't yet decoupled OBAKit from the Stripe SDK, and so I have resorted to `canImport()` macros to help me understand if we're in an environment that has linked in the Stripe SDK or not.

Down the road, we should decouple the Stripe SDK from OBAKit and move the dependency into the app layer, just like Firebase and OneSignal. But, for now, expediency is key, and so we'll just do what works.